+++ setupTests/enzyme.js
@@ -1,0 +1,4 @@
+const Adapter = require('enzyme-adapter-react-16');
+const enzyme = require('enzyme');
+
+enzyme.configure({ adapter: new Adapter() });